requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
PyMySQL==1.1.0
ReportLab==4.0.4
Jinja2==3.1.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
